requirements-web.txt
Flask>=3.1.0,<4
huggingface_hub>=0.31.0,<1
Pillow>=10.4.0,<12
